+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Braintree Android SDK Release Notes
 
+## unreleased
+
+* GooglePay
+  * Add optional `resultKey` parameter to `GooglePayLauncher` constructors, allowing a unique registration key per launcher instance
+* ThreeDSecure
+  * Add optional `resultKey` parameter to `ThreeDSecureLauncher` constructors, allowing a unique registration key per launcher instance 
+  
 ## 5.32.0 (2026-08-27)
 
 * GooglePay